roles/was/files/app/requirements.txt
fastapi
uvicorn
pyjwt
python-multipart
redis>=5.0.0
sqlalchemy
psycopg2-binary
httpx
asyncpg
minio